support/requirements.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.11.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==3.9.2
flake8-bugbear==21.11.29
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-isort==4.1.1
flake8-polyfill==1.0.2
flake8-variables-names==0.0.4
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
pep8-naming==0.12.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
redis==4.1.1
requests==2.27.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
testfixtures==6.18.3
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
